+++ KFXAdditions/Classes/Foundation/NSNumber/NSNumber+KFXAdditions.m
@@ -22,12 +22,23 @@
 	return [self kfx_isLessThanNumber:anotherNumber] ? self : anotherNumber;
 }
 
+/// Returns the lowest number from the array of numbers
++(NSNumber *)kfx_lowestNumberFromNumbers:(NSArray<NSNumber*>*)otherNumbers{
+    
+    NSArray *sorted = [otherNumbers sortedArrayUsingSelector:@selector(compare:)];
+    return sorted.firstObject;
+}
 
 /// Returns the highest number out of the receiver and anotherNumber
 - (NSNumber *)kfx_highestNumber:(NSNumber *)anotherNumber{
 	return [self kfx_isGreaterThanNumber:anotherNumber] ? self : anotherNumber;
 }
 
+/// Returns the highest number from the array of numbers
++(NSNumber *)kfx_highestNumberFromNumbers:(NSArray<NSNumber*>*)otherNumbers{
+    NSArray *sorted = [otherNumbers sortedArrayUsingSelector:@selector(compare:)];
+    return sorted.lastObject;
+}
 
 /// Returns YES if the receiver == anotherNumber
 -(BOOL)kfx_isEqualToNumber:(NSNumber *)anotherNumber{