+++ KFXAdditions/Classes/Foundation/Strings/NSAttributedString+KFXAdditions.h
@@ -16,14 +16,42 @@
 //--------------------------------------------------------
 #pragma mark - Convience Initilisers
 //--------------------------------------------------------
+/// Returns an NSAttributedString with the given string and no attributes
 +(instancetype)kfx_instanceWithString:(NSString*)string;
+
+/// Returns an NSAttributedString with the given string and attributes
 +(instancetype)kfx_instanceWithString:(NSString*)string attributes:(NSDictionary*)attributes;
+
+/// Returns an NSAttributedString with the given format string and no attributes
 +(instancetype)kfx_instanceWithFormat:(NSString*)format,...;
-+(instancetype)kfx_instanceWithAttributes:(NSDictionary*)attributes format:(NSString*)format,...;
+
+/// Returns an NSAttributedString with the given format string and the attributes
++(instancetype)kfx_instanceWithAttributes:(NSDictionary*)attributes
+                                   format:(NSString*)format,...;
+
+/// Returns an NSAttributedString with the given string and the attributes
++(instancetype)kfx_instanceWithString:(NSString*)string
+                                 font:(UIFont*)font
+                               colour:(UIColor*)colour;
+
+/// Returns an NSAttributedString with the given string and the attributes
++(instancetype)kfx_instanceWithString:(NSString*)string
+                                 font:(UIFont*)font
+                               colour:(UIColor*)colour
+                            alignment:(NSTextAlignment)alignment;
+
+/// Returns an NSAttributedString with the given string and the attributes
++(instancetype)kfx_instanceWithString:(NSString*)string
+                                 font:(UIFont*)font
+                               colour:(UIColor*)colour
+                            alignment:(NSTextAlignment)alignment
+                          lineSpacing:(CGFloat)lineSpacing;
+
 
 //--------------------------------------------------------
 #pragma mark - Ranges
 //--------------------------------------------------------
+/// Returns the NSRange for the receiver's string
 -(NSRange)kfx_rangeOfString;
 
 @end