+++ KFXAdditions/Classes/Foundation/Strings/NSAttributedString+KFXAdditions.m
@@ -49,8 +49,66 @@
         va_end(args);
     }
     return [self kfx_instanceWithString:string attributes:attributes];
+}
+
+/// Returns an NSAttributedString with the given string and the attributes
++(instancetype)kfx_instanceWithString:(NSString*)string
+                                 font:(UIFont*)font
+                               colour:(UIColor*)colour{
+ 
+    if (string == nil || font == nil || colour == nil) {
+        return nil;
+    }
+    NSDictionary *attributes = @{
+                                 NSFontAttributeName:font,
+                                 NSForegroundColorAttributeName:colour                                 };
+    NSAttributedString *attString = [self kfx_instanceWithString:string attributes:attributes];
+    return attString;
+}
+
+/// Returns an NSAttributedString with the given string and the attributes
++(instancetype)kfx_instanceWithString:(NSString*)string
+                                 font:(UIFont*)font
+                               colour:(UIColor*)colour
+                            alignment:(NSTextAlignment)alignment{
+    
+    if (string == nil || font == nil || colour == nil) {
+        return nil;
+    }
+    NSMutableParagraphStyle *paraStyle = [[NSMutableParagraphStyle alloc]init];
+    paraStyle.alignment = alignment;
+    NSDictionary *attributes = @{
+                                 NSFontAttributeName:font,
+                                 NSForegroundColorAttributeName:colour,
+                                 NSParagraphStyleAttributeName:paraStyle
+                                 };
+    NSAttributedString *attString = [self kfx_instanceWithString:string attributes:attributes];
+    return attString;
+}
+
+/// Returns an NSAttributedString with the given string and the attributes
++(instancetype)kfx_instanceWithString:(NSString*)string
+                                 font:(UIFont*)font
+                               colour:(UIColor*)colour
+                            alignment:(NSTextAlignment)alignment
+                          lineSpacing:(CGFloat)lineSpacing{
+    
+    if (string == nil || font == nil || colour == nil) {
+        return nil;
+    }
+    NSMutableParagraphStyle *paraStyle = [[NSMutableParagraphStyle alloc]init];
+    paraStyle.lineSpacing = lineSpacing;
+    paraStyle.alignment = alignment;
+    NSDictionary *attributes = @{
+                                 NSFontAttributeName:font,
+                                 NSForegroundColorAttributeName:colour,
+                                 NSParagraphStyleAttributeName:paraStyle
+                                 };
+    NSAttributedString *attString = [self kfx_instanceWithString:string attributes:attributes];
+    return attString;
 
 }
+
 
 //--------------------------------------------------------
 #pragma mark - Ranges